@@ -7,13 +7,13 @@
     "lib"
   ],
   "scripts": {
-    "autod": "autod --check",
+    "autod": "autod",
     "pkgfiles": "egg-bin pkgfiles --check",
     "lint": "eslint --ext .js lib test",
     "test": "npm run lint && npm run test-local",
     "test-local": "egg-bin test",
     "cov": "egg-bin cov",
-    "ci": "npm run autod && npm run lint && npm run cov"
+    "ci": "npm run autod -- --check && npm run lint && npm run cov"
   },
   "repository": {
     "type": "git",
@@ -30,24 +30,24 @@
   },
   "homepage": "https://github.com/node-modules/serialize-json#readme",
   "dependencies": {
-    "debug": "^3.0.1",
-    "is-type-of": "^1.2.0",
-    "utility": "^1.12.0"
+    "debug": "^3.2.6",
+    "is-type-of": "^1.2.1",
+    "utility": "^1.15.0"
   },
   "devDependencies": {
-    "autod": "^2.9.0",
+    "autod": "^3.0.1",
     "beautify-benchmark": "^0.2.4",
     "benchmark": "^2.1.4",
     "egg-bin": "^1.11.1",
-    "egg-ci": "^1.8.0",
-    "eslint": "^4.7.2",
-    "eslint-config-egg": "^5.1.1",
-    "mm": "^2.2.0"
+    "egg-ci": "^1.10.0",
+    "eslint": "^4.19.1",
+    "eslint-config-egg": "^6.0.0",
+    "mm": "^2.4.1"
   },
   "engines": {
     "node": ">= 4.0.0"
   },
   "ci": {
-    "version": "4, 6, 8"
+    "version": "4, 6, 8, 10, 11"
   }
 }
