@@ -26,7 +26,11 @@ class JSONDecoder {
   }
 
   _decodeString(str) {
-    return str.replace(REG_STR_REPLACER, a => DECODER_REPLACER[a]);
+    // avoid Parent in (sliced string)
+    // https://github.com/nodejs/help/issues/711
+    // https://stackoverflow.com/questions/31712808/how-to-force-javascript-to-deep-copy-a-string
+    const r = str.replace(REG_STR_REPLACER, a => DECODER_REPLACER[a]);
+    return (' ' + r).slice(1);
   }
 
   _decodeDate(str) {
@@ -81,11 +85,11 @@ class JSONDecoder {
           const token = this.tokens[this.tokensIndex];
           if (token === ']') {
             debug('--> unpack buffer end, %j', arr);
-            return new Buffer(arr);
+            return Buffer.from(arr);
           }
           arr.push(this._unpack());
         }
-        return new Buffer(arr);
+        return Buffer.from(arr);
       }
       case '#': // error
       {
