@@ -40,7 +40,7 @@ class JSONEncoder {
     packed += `^${this._pack(ast)}`;
 
     debug('pack the json => %s', packed);
-    return new Buffer(packed);
+    return Buffer.from(packed);
   }
 
   _pack(ast) {
